Another, smaller, overhaul.
Treat Hostname lines in the config file as just another Host pattern and
not as a canonical hostname (that was causing IP addresses to replace
more identifiable Host pattern entries).

Consolidate the logic for handling new entries into a helper function
(and make it the same for both config and known_hosts).

Add a few new tests and update some tests to match the above changes.

Move username styling to parse_config to avoid complications in seen
checking, etc.